multimodule project created
