+++ pokemon-persistence/pom.xml
@@ -4,13 +4,13 @@
 
     <parent>
         <groupId>cz.muni.fi.pa165</groupId>
-        <artifactId>eshop-parent</artifactId>
+        <artifactId>pokemon-parent</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>pokemon-persistence</artifactId>
+    <artifactId>pokemon-persistance</artifactId>
     <packaging>jar</packaging>
-    <name>Persistence Layer and Beans Validation</name>
+    <name>Pokemon persistance</name>
 
     <dependencies>
         <dependency>
@@ -112,6 +112,13 @@
             <groupId>org.testng</groupId>
             <artifactId>testng</artifactId>
             <version>6.14.3</version>
+            <scope>test</scope>
+            <type>jar</type>
+        </dependency>
+        <dependency>
+            <groupId>javax.inject</groupId>
+            <artifactId>javax.inject</artifactId>
+            <version>1</version>
             <scope>test</scope>
             <type>jar</type>
         </dependency>
