+++ pokemon-persistence/src/main/java/cz/muni/fi/pa165/pokemon/league/participation/manager/dao/BadgeDAOImpl.java
