+++ pokemon-persistence/src/test/java/cz/muni/fi/pa165/pokemon/league/participation/manager/dao/TrainerDAOTest.java
